Struct snapshot::Snapshot

pub struct Snapshot {
    pub(crate) hdr: SnapshotHdr,
    pub(crate) version_map: VersionMap,
    pub(crate) target_version: u16,
}

The Snapshot API manages serialization and deserialization of collections of objects that implement the Versionize trait.

Fields
hdr: SnapshotHdrversion_map: VersionMaptarget_version: u16
Implementations

impl Snapshot

pub fn new(version_map: VersionMap, target_version: u16) -> Snapshot

Creates a new instance which can only be used to save a new snapshot.

pub fn get_data_version<T>( reader: &mut T, version_map: &VersionMap ) -> Result<u16, Error>
where T: Read + Debug,

Fetches snapshot data version.

pub fn unchecked_load<T: Read + Debug, O: Versionize + Debug>( reader: &mut T, version_map: VersionMap ) -> Result<(O, u16), Error>

Attempts to load an existing snapshot without CRC validation.

pub fn load<T: Read + Debug, O: Versionize + Debug>( reader: &mut T, snapshot_len: usize, version_map: VersionMap ) -> Result<(O, u16), Error>

Attempts to load an existing snapshot and validate CRC.

pub fn save<T, O>(&mut self, writer: &mut T, object: &O) -> Result<(), Error>
where T: Write + Debug, O: Versionize + Debug,

Saves a snapshot and include a CRC64 checksum.

pub fn save_without_crc<T, O>( &mut self, writer: &mut T, object: &O ) -> Result<(), Error>
where T: Write, O: Versionize + Debug,

Save a snapshot with no CRC64 checksum included.

pub(crate) fn format_version_map() -> VersionMap